Introduce RetryCommandDispatcher
